#Classification_task/requirements.txt
torch>=1.12
torchvision>=0.13
scikit-learn>=1.0
xgboost>=2.1
joblib>=1.2
pandas>=1.4
numpy>=1.21
matplotlib>=3.5
streamlit>=1.18
Pillow>=9.0
tqdm>=4.64